requirements.txt
boto3==1.4.2
botocore==1.4.85
chalice==0.5.0
click==6.2
docutils==0.13.1
funcsigs==1.0.2
futures==3.0.5
jmespath==0.9.0
mock==2.0.0
nose==1.3.7
pbr==1.10.0
python-dateutil==2.6.0
python-dotenv==0.6.1
PyXB==1.2.5
s3transfer==0.1.9
six==1.10.0
typing==3.5.2.2
virtualenv==15.1.0
